+++ Jeopardy/src/app/app.module.ts
@@ -7,7 +7,7 @@ import { StartComponent } from './start/start.component';
 import { TeamComponent } from './team/team.component';
 import { NavBarComponent } from './nav-bar/nav-bar.component';
 import { HomeComponent } from './home/home.component';
-import { LeaderboardComponent } from './leaderboard/leaderboard.component';
+import { LeaderboardComponent } from './leaderboard_components/leaderboard/leaderboard.component';
 import { GameboardComponent } from './gameboard/gameboard.component';
 import { FormsModule } from '@angular/forms';
 import { CategoryComponent } from './category/category.component';
@@ -17,7 +17,7 @@ import { AdminComponent } from './admin/admin.component';
 import { CreateComponent } from './create/create.component';
 import { StatusbarComponent } from './statusbar/statusbar.component';
 import { SavedgamesComponent } from './savedgames/savedgames.component';
-import { PlayersComponent } from './players/players.component';
+import { PlayersComponent } from './leaderboard_components/players/players.component';
 
 @NgModule({
   declarations: [
