+++ Jeopardy/src/app/leaderboard_components/leaderboard/leaderboard.component.css
